My last lame attempt
This is the point where my code was when I gave up last night and went
to the subreddit for help.  Looking at jimcasey's code I see this was
never going to work because I wasn't handling the rotations properly.

Also, making triangles with the two nearest points and using their
areas as an invariant is a really clever idea. I was getting there
with my initial pairwise distance idea, but this way is better.